# swagger_files/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "myBuzzn App API",
        "description": "An app to investigate your power consumption habits for BUZZN customers."
    },
    "paths": {
...
        "/individual-disaggregation": {
            "get": {
                "summary": "Shows the power curve disaggregation from the given starting point for a user.",
                "description": "",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "integer"
                                }
                            },
                            "example": {
                                "2020-01-15 10:01:10": {
                                    "Waschmaschine-1": 0,
                                    "Sp\u00fclmaschine-1": 0,
                                    "Durchlauferhitzer-2": 0,
                                    "Durchlauferhitzer-3": 0,
                                    "Grundlast-1": 2500000,
                                    "Durchlauferhitzer-1": 0
                                }
                            }
                        }
                    },
                    "206": {
                        "description": "No history available"
                    }
                },
                "parameters": [
                    {
                        "name": "begin",
                        "in": "query",
                        "description": "Start time of disaggregation as unix timestamp. Default is two days ago at 00:00:00 am.",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "disaggregation"
                ],
                "produces": [
                    "application/json"
                ],
                "operationId": "individualDisaggregation"
            }
        },
...
        "/group-disaggregation": {
            "get": {
                "summary": "Shows the power curve disaggregation from the given starting point for a  group.",
                "description": "",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "integer"
                                }
                            },
                            "example": {
                                "2020-01-15 10:01:10": {
                                    "Waschmaschine-1": 0,
                                    "Sp\u00fclmaschine-1": 0,
                                    "Durchlauferhitzer-2": 0,
                                    "Durchlauferhitzer-3": 0,
                                    "Grundlast-1": 2500000,
                                    "Durchlauferhitzer-1": 0
                                }
                            }
                        }
                    },
                    "206": {
                        "description": "No history available"
                    }
                },
                "parameters": [
                    {
                        "name": "begin",
                        "in": "query",
                        "description": "Start time of disaggregation as unix timestamp. Default is two days ago at 00:00:00 am.",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "disaggregation"
                ],
                "produces": [
                    "application/json"
                ],
                "operationId": "groupDisaggregation"
            }
        },
        "/set-password": {
            "post": {
                "summary": "Sets an initial password. Does not require the user to be logged in.",
                "description": "",
                "responses": {
                    "200": {
                        "description": "password passes the validation"
                    },
                    "400": {
                        "description": "password fails the validation or the token is invalid",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "errorName": {
                                    "type": "string"
                                },
                                "errorDescription": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "password to set and user token",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "password": {
                                    "type": "string"
                                },
                                "token": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "tags": [
                    "profile"
                ],
                "consumes": [
                    "application/json"
                ],
                "operationId": "setPassword"
            }
        },
...
    },
...
}